Parsing incoming JSON without any prior knowledge of data structure
The incoming data structure will never be known explicitly. For example an Ethereum ABI will have a variety of unknown key value pairs. In Rust, as a rule, you create a struct which defines the data and then you parse the string into JSON based directly on the variables in the struct. This would never work in our situation.
We now have the ability to parse random JSON input and only call what we want to using this notation `json_object["the_key_we_want"]`. This is extremely useful.
Also we are not accepting unlimited strings as input into the service call. That is only allowed in debug mode because it is a DDOS vector. Instead we accept a Data object in the format of  bytes vec and then we peek into that using the proper APIs.